+++ src/main/java/common/HttpStatus.java
@@ -1,0 +1,23 @@
+package common;
+
+public enum HttpStatus {
+    OK(200, "OK"),
+    NOT_FOUND(404, "NOT FOUND");
+
+    private final int code;
+    private final String message;
+
+    HttpStatus(int code, String message) {
+        this.code = code;
+        this.message = message;
+    }
+
+    public int getCode() {
+        return code;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+}
